v-show directive AND v-if/v-else VS v-show
v-show="some condition like in the if statement"

if the condition is true, the element is shown.

The difference is that, v-if and v-else ADD or REMOVE elements from the DOM, which costs performance.

The v-show directive uses CSS, to hide or unhide the element. Which saves performance but keeps a bunch of elemnts that we don't currently need.

VERDICT:

Use v-if and v-else typically.

Fall back to v-show by your own judgement, for elements who have their visibility change a lot.