v-for - Iterating Objects
While v-for is mostly used for iterationg through lists, it can also be used to iterate through objects.

Syntax:

<ul>
        <li v-for="(value, key) in {name: 'Max', age: 31, occupation: 'CS related
          something'}"> {{key}}: {{value}} </li>

</ul>

This will iterate on the object, and create an <li> tag for each key.

When iterating through objects, the 2nd argument you can fetch is the key (name, age, occupation) rather than the index (d'uh)