v-for - ALWAYS use keys! AND @click.stop modifier
key attribute is a Vue specific atrribute.

Its almost a must to use :key whenever you are using v-for.

v-for="..." :key="a-unique-identifier" (or v-bind:key="a-unique-identifier")

we can use the @click.stop modifier under a tag which is listening for @click, to stop the click behaviour/listening on the specific child tag. In the code, we use it to stop the @click behaviour when the user clicks on the input tag to input text.

Also, notice how in our code, we dont use "index" as the :key. Thats because the index is not unique for each <li> element. Infact, 0 is always the index of the first <li> element 1 is always the index of the second <li> element and so on...

To further explain this: Let's say we have 2 elements in a list;

1st element has index 0,
2nd element has index 1.

If we delete the 1st element, the 2nd element will now have index 0. Hence proven XD.

The best thing to use as :key: is the id from a database.