Receiving Props in a different/better way + IMPORTANT CONCEPTS
For each prop, we can setup extra configuration. Example, lets take a prop 'name'

We define the configuration in an object.

name: {
 type: String,
 required: true,
 default: 'empty',
 validator: function () { }
}

- type sets the type we expect
- required sets that the prop is required or not
- default sets the default value if the prop is not provided (if required is set to false),
  we can also call a function here if the value is not provided.
- validator holds a function which should return true or false.

To see a good example of using default and validator, ofcourse, check the code.

NOTE: 'required: true' will not throw an error and cause the site to crash, it just instead helps us as a developer to spot an error quicker and help with debugging.

-----
!! I M P O R T A N T !!
-----
To send anything other than a string in the HTML tag from App.vue, we have to bind it using v-bind: or the shorthand semi-colon : . This is very logical and basic, but I know I will forget, so this needs to go here.

Using this, we can use the <friend-contact> component tag only once, and dynamically render it with different values using v-for. (DON'T FORGET TO USE :key :key:)