Delete a friend AND .filter() javascript function
array.filter() is a vanilla javascript function that returns a new array based on the array on which you're calling it, except for the elements which you want to filter out.

It accepts a function, the function accepts the current element of the array which is being iterated, and returns true or false for each element. When it returns true, the element is added to the new array, and if it returns false then the element is not added. We then assign the resultant array to this.friends.

For example, if 'friends' is the name of the array, you can do:

this.friends = this.friends.filter( (dummyNameForOneElement) => { if this return true else return false }

this.friends now contains all the elements except for which the function returned false.

In the code, the friends array now contains all the elements, except the friend with the ID the FriendConact component sent.