The problem with the last commit was REPITITION
We saw repitition in data; topics used twice, once in data and once in provide.

To tackle that, we can use the provide function in another way. We can instead use the provide option as a function that returns the provide data. Thanks to that, we can now use this.topics and reach out to our data property and just set it equal to our provide 'key'.