Dynamic Components - <component> element
the componnet element is provided by vue, and it doesnt show anything by itself.

It needs one key property; i.e: the 'is' property.

the is property tells component which component of all the components you defined should be shown in place ofthis component.

<component is="active-goals"></components>

As you may have guessed, we can bind the is property and call methods off of it, so we can render multiple/different components under a given condition.