<keep-alive> element
It does what the name suggests. It keeps the componenet alive.

Say that you have an input field in the ManageGoals.vue file. If you save and go to the website, and input something without pressing enter, and then click on active goals, the text that you entered will be lost.

Thats because vue destroys the old component from the DOM and reintroduces it when you press ManageGoals.

This is the default behaviour. Y ou acn wrap the <component> tag in <keep-alive> to prevent this.