Teleporting elements
Wrap the to-be moved HTML element in the <teleport> element.

teleport expects an argument: to=""

<teleport to="<css-selector-here>">

We used it  in our code here to solve the HTML nesting problem. The error-alert component we were spawning on an invalid input was deeply nested in our HTML. We therefore wrapped the error-alert in teleport and teleported it to <body>