Made multiple components and slots + one important approach
Made a dynamic button component the BaseButton

Demonstrated a very good approach to making dynamic components, which are particulary helpful when working with a team.

I am just calling base-button wherever I want a button, where in I can pass a prop (not a vue bound prop) called "mode" which can be set to any name you and your team picks up, for which the styling exists in the component itself, you just pick the styling you prefer in different components. This has reduced the CSS by a ton. And is super clean