v-model works on <select> element just like it does on the <input> element
If you have a group of checkboxes or radio buttons, that share the same name="" atrribute, then you get an array where all the checked values will be elements in.

So make sure to bind each input of type="checkbox" with the same data property (which should be an array, in our code, its the "interest" data property).

Give each input different value="" properties to identify them easily and to make sure that all of the checkboxees dont become unchecked when you uncheck one of them.

If you have a single checkbox, then you get a true or false.