Form Validation | @blur event
You can use the @blur event listener which Vue provides.

@blur basically watches out and triggers a method once a user focuses and then defocuses the element, in our case, the input element (the text field).

And in our code, we are firing the validateInput() method. Which validates if the input field was empty or not.

This is a basic example of form validation, and it can be used to any extent to validate any kind of input.

-- Dynamic :class attribute to implement class attribute on a condition --

In this particular commit, we also added :class attribute, to only show the paragraph below the username input field if the usernameValidity was invalid. this was added using:

:class="{<classname>: <some-data-property> ==/</>/!= <some-value-or-string}"

So,

:class="{invalid: usnernameValidity == 'invalid'} "