fetch() and then() and promsies and Arrow function + this
--- Using Fetch to GET data ---
loadExperiences() {
      fetch('https://vue-http-demo-449f9-default-rtdb.firebaseio.com/surveys.json');
        // method: 'GET', // the GET method is the default, so we don't need to set this.
        // headers: { // We dont need the headers aswell, since we have no body.
        //   'Content-Type': 'application/json' // Telling the server that we'll add some data to this request, whhich will be in the JSON format.
        // },
        // Since we're getting data, we also dont need to set any body. Because a get request doesnt carry any data. \/

fetch() returns an object on which we can listen for the data once it is there, to set up some code which will be executed when that data is there.

--- then() method ---

We can add such a listener, by adding the then() method on the result of fetch. The then method  takes a function which will be executed once the data is there. And in this function, we will get an argument automatically which is the "response" argument.

fetch('https://.....").then(function(response) {
   // code to execute here
});

--- response ---

 if (response.ok) { // response has an ok property which indicates that the request was successful or not. If its ok, then we can work on the data:
        // response has a various methods for that;
        // (1) .json() which will parse the data thats part of the response if its in JSOn format.
          return response.json();
        }

--- Another then() after the first then() ---

...
return response.json();
        }
      }).then(); // Another then method after that first then(). This one will be trigerred when this return promise is done.

The argument this then() receives is the data that the response.json yielded. Kind of common sense.

--- Arrow function and Vue ---

line 62: this.results = results; // setting the Vue "results" variable equal to our temporary results variable from line 52. For this line to work, you have to make sure to use Arrow functions, rather than the function() {} syntax. BECAUSE, if you use the normal function() syntax, the meaning of the "this." keyword changes. If you use arrow functions on the other hand,  "this." means the same thing inside as outside of the function.