Server Side Error Handling
NOTE: All errors are fixed, the code is working. Generate fake errors for practice by using the information below.:

What if Firebase returns an error that the data couldn't be fetched?
The catch() statement doesnt catch that. So...

Fake Error # 1: Remove .json from the URL >>>
Error # 1 can be handled by a simple catch() statement.

Fake Error # 2: Remove JSON.stringify from the body >>>
To handle errors of kind Error # 2, we need to use then(), which is alawys triggered and then:

if (response.ok) {
          // do something ...
        } else {
          throw new Error('Could not save data!') // This throws a new error, and then automatically reaches the end block below.
        }