Another tag that ships with vue-router
The first one was <router-view> which allowed us to tell vue-router where to lead the components.

Now we're looking at <router-link>:

Its a little bit like the built in anchor tag <a></a>

A special tag which will not load a different page and reload the entire app, losing the current state.

Instead, it's a tag which prevents the browser default and doesn't reload, analyzes the link the user clicked and then loads the appropriate component and updates the URL.

router link needs to be configured, by using the to="" property.

Which takes various different values, and the simplest way of using it is to provide a string to it which holds the path you wanna go to. It should be one of the paths you are supporting in your main.js

We can use a{} (the anchor tag) to style the router-link

If you inspect the <router-link> tags in the browser, you'll see that they are actually replace with <a> tags by Vue. And also, Vue adds a CSS class to the currently selected <router-link>. This helps us to provide styling to the currently selected tag. Nifty. We'll see this next.

Docs:
router-link: https://router.vuejs.org/api/#router-link
router-view: https://router.vuejs.org/api/#router-view