<router-link> active classes AND programmatically redirecting using this.$router
Two classes are automatically added to the currently active <router-link> tag.

these are
1) "router-link-active"
2) "router-link-exact-active"

We can use the first one, to style all active nested tags aswell.

So if we had url.com/teams/teamA
the effect will still be in effect.

On the other hand if we use the exact-active class, it will only work if we have url.com/teams only.

------------------
We can provide additional configuration to our createRouter in main.js, relating to the classes that are added to the active <router-link> element. These are the linkActiveClass and linkExactActiveClass as the key, and the value should be a string which you instead need to use as the CSS class name.

------------ this.$router ---------------
Since we are using the vue-router, we have a special property available, this.$router('/path'):

FUN NOTE: This is the same methods thats called behind the scenes when you click a <router-link>, so clicking routerlink and using this.$router.push() means the same thing.

this.$router.push('/teams'); // Addind a new route to this routing memory the browser keeps. We'll be now redirected programaticalyl to /teams.

There are other functions this.$router has to emulate different browser behaviour. Like the back() and forward() to go back a webpage or one up.

this.$router.back()
this.$router.forward()

READ THESE DOCS!