A new problem - created() hook and the $route - new <router-link> doesnt work.
We will now create a new problem for ourselves. To learn.

We were using our dynamic loading logic in the created() hook.

If we introduce a button, which takes us to team 2, so..
<router-link to="/teams/t2"></router-link> and click on it after any other route has already been called.

So like in the /teams/t1 page, it won't work. The URL will change, but the component will not be loaded with new data. So to fix this, we have to use watchers.

We will watch for the $route property.

-- Note about $route: Whenever the URL changes, one important thing changes, the $route changes. It holds the latest information about the loaded route. So this $route will always hold the latest parameter for example. --

THINK ABOUT IT, ITS COMMON SENSE, BUT USEFUL TO KNOW!