Update main.js
We can also use the alias option.

Difference in alias and redirect.

Redirect is used on a different route. So if I want to redirect to a different path/route, then I will have to make another route element of the routes array, and then redirect it to a pre-existing route.

Alias is instead added to the existing route where we wish for the same component to be loaded when the user is on the "path" or the "alias". ALSO A MAJOR DIFFERENCE is, when we define a path as alias, then we will not see the URL change to its "path" counterpart, instead, the alias path will be used as the URL.

So in our use case, it's better to use redirect. Consequently, alias will be removed after this commit.