Rendering multiple routes with named Router Views
In the route config, we can add an option for components instead of component.

components now take an object, which has key value pairs.

keys identify different router views, and the values are the components that should be loaded. "default" corresponds to unnamed router-view.

Any route with just 'component' (without the 's') will be caught by the unname route. So if we enter an invalid URL, we'll only be shown the 'NotFound' Component.