Navigation Guards
Preventing to access a certain route. To ensure a user doesn't accidentally navigate away from a page where user has unsaved data etc.

Guards are functions that are called automatically by Vue router when a page changes, or to be precise, when a navigation action started.

router.beforeEach( (to, from, next) => {

 beforeEach takes a function as an arugment, which will be called by the vue router, whenever we navigate from one page to another. Before Each navigation action, this function will be called. This function takes 3 arguments. "next" is a function we have to call to whether confirm or cancel this navigation action. This part is what makes it a 'guard'.

next(); // If we call next() like this, we allow all navigations.

next(false) // stops all navigation.

We can also pass a string with a route we wanna navigate to (one of the ones we registered above) OR such a navigation object.

if (to.name === 'team-members' ) {
     next();
   } else {
     next({name: 'team-members', params: { teamID: 't2'} });
   }

The logic above makes no sense, its just to explain how it works.