Another Global Guard
The afterEach( () => {}); gaurd. Which expects a function, which in turn accepts to and from parameters.

afterEach will only run once after the navigation has been confirmed. So in here, we can't deny or anything. Thats why it doesnt accept the next param ofcourse.

It's used to send analytics data etc. We dont use it to control what the users see on their screen.