Utilizing Route Metadata
...
      path: '/teams',
      meta: {needsAuth: true},
      components: { default: TeamsList, footer: TeamsFooter },
...

meta property holds any kind of values in it. This is useful because we can access this meta field in every place where the $route object is available. So we can use it inside components, as well as Navigation Guards. We can now use it in line 59 of main.js to check if the user is authenticated or not.

... we can now use this for example, like this:

  if (to.meta.needsAuth) {
    console.log('Needs Auth!');
  } else {
    next();
  }