Mutations
The previous way of implementing Vuex is not ideal. As we can introduce errors.

We should use it in a different way. So we'll not directly co

Mutations are clearly defined methods which have the logic to update the states.
Then from inside of our components, we should just trigger those mutations instead of directly manipulating the state.

Now every components that wants to edit a state, will do in the same way as described in the mutation. So in this way, we..

1) Eliminate a chances at errors (one such error is simulated in line 27 of App.vue)

2) Don't have to retype or reintroduce logic for updating/mutating/changing that state in every component.

------ Defining Mutations in main.js --------
We define our mutations in main.js of course. Alongside state() we also have a mutations: {} option. Here we define methods that will update/mutate our state.

And as such each method AUTOMATICALLY accepts a "state" argument which can be used to access the states defined above. Pretty simple huh?

------ Calling/Triggering mutations in components -----

this.$store.commit('increment'); // Calling the mutation (use this line in a method etc)

$store also has a commit function which takes in a string as an argument.
In the code above, we provide the name of the mutation created in main.js as a string to the commit function.