Getters to get data - A better way
Getters are computed properties defined directly in the store.

Just like updating/mutating state was a bad practice directly in a component, getting data or using computed property inside a component to point at a state is also a bad practice.

We instead use getters. And now, in the computed properties of a component, we instead:

return this.$store.getters.finalCounter

---- Defining Getters ----
We define getters using:

finalCounter(state) {
      return state.counter * 2;
},

normalizedCounter(_, getters) {
      const finalCounter = getters.finalCounter;
      if (finalCounter < 0) {
        return 0;
      }
      if (finalCounter > 100) {
        return 100;
      }
      return finalCounter;
}