Running Async code with Actions
Mutations do not allow Asynchronous code.

So actions is just another option alongside mutations and getters. It can have the same name as the mutations. Actions are just basically used to commit the mutations.

How do we commit inside an action method?
Each action method receives a context object as argument (It also receives the payload as an argument if passed. Now context has the commit method. Its the same syntax and same everything. We just pass in the name of the mutation.

increment(context, payload) {
   setTimeout( () => {
      context.commit('increment', payload); // just forwarding the payload, not accessing the value.
   }, 2000);

But we can write asynchronous code in actions, which if used in mutations can lead to errors and bugs.

Now to call actions from components instead of directly committing them, we use:

this.$store.dispatch('increment', payload)