mapGetters and mapActions
---- mapGetters ----
Instead of making our own computed element for each getter we have in our main.js store. We can:

1) Import mapGetters (and mapActions)

import {mapGetters} from 'vuex'

2) Use the spread operator and use the mapGetters method

...mapGetters(['finalizedCounter', 'normalizedCounter']}

mapGetters takes in an array of strings, each string corresponds to a getter name.

---- mapActions ----
mapActions works the same way, we import, sue the spread operator to merge with the methods object. It takes an array where we define all the action names which we want to use in our component.

mapActions makes the methods, calls dispatch in them and everything. Its equivalent to what we wrote before. (See code)

...mapActions(['increment', 'decrement'])

What about the payload then?

To send the payload, we can just pass the payload as an argument in the <template> section where we call the action.

mapMethods also exists. And its simple to use just like mapActions.